Displaying logs have been moved to dedicated step. No more side effects.
