+++ packages/ckeditor5-dev-env/lib/release-tools/tasks/generatechangelogforsubpackages.js
@@ -12,6 +12,7 @@ const getNewReleaseType = require( '../utils/getnewreleasetype' );
 const cli = require( '../utils/cli' );
 const versionUtils = require( '../utils/versions' );
 const changelogUtils = require( '../utils/changelog' );
+const displayCommits = require( '../utils/displaycommits' );
 const displaySkippedPackages = require( '../utils/displayskippedpackages' );
 const displayGeneratedChangelogs = require( '../utils/displaygeneratedchangelogs' );
 const executeOnPackages = require( '../utils/executeonpackages' );
@@ -80,8 +81,10 @@ module.exports = function generateChangelogForSubPackages( options ) {
 		log.info( chalk.bold.blue( `Generating changelog for "${ dependencyName }"...` ) );
 
 		return getNewReleaseType( transformCommitFunction, { tagName } )
-			.then( response => {
-				const newReleaseType = response.releaseType !== 'skip' ? response.releaseType : null;
+			.then( result => {
+				displayCommits( result.commits );
+
+				const newReleaseType = result.releaseType !== 'skip' ? result.releaseType : null;
 
 				return cli.provideVersion( packageJson.version, newReleaseType );
 			} )
